+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Fixed
+- Removed spurious start_gap and end_gap in TraceTransition.
+
 ## [0.4.2] - 2021-04-16
 ### Changed
 - SegmentList.draw() now returns a list of the drawn structures.